Add support for configurable secondary icons
Adds a preference and menu item to turn secondary icons on and off. They are on by default. When they are turned on the existing behavior remains unchanged. When they are turned off each file will display only the most specific icon relevant. i.e. A `gulpfile.js` will show the Gulp logo and not the JS logo but `double.php.js` will show only the JS logo and not the PHP one.