Bug fix and change in icon behaviour
This commit changes the icon matching behaviour to only look at the last two elements of a file name and give more importantce to the default icon when the extension is not known. This commit also fixes a bug where an unknown file extension with no prefix could cause an infinite loop. As a side effect this commit also makes a big performace improvement as the loop is gone.

This fixes #89 and fixes #88